requirements.txt
fastapi
uvicorn
sqlalchemy
pymysql
cryptography
pydantic
python-dotenv
alembic
cors
python-multipart
# Google Drive API
google-api-python-client
google-auth
google-auth-httplib2
google-auth-oauthlib
cloudinary
pytz
